fix: MinigameOverlay, QuizReadyScreen
