+++ app/src/main/java/at/aau/serg/websocketbrokerdemo/ui/theme/minigames/quiz/QuizReadyScreen.kt
@@ -1,0 +1,102 @@
+package at.aau.serg.websocketbrokerdemo.ui.theme.minigames.quiz
+
+import android.graphics.BitmapFactory
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.asImageBitmap
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+fun QuizReadyScreen(
+    onReadyClick: () -> Unit
+) {
+    val context = LocalContext.current
+    val foxBitmap = remember {
+        context.assets.open("baron_fox.png").use { inputStream ->
+            BitmapFactory.decodeStream(inputStream).asImageBitmap()
+        }
+    }
+
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(Color.Black.copy(alpha = 0.7f)),
+        contentAlignment = Alignment.Center
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth(0.6f)
+                .clip(RoundedCornerShape(16.dp))
+                .background(Color(0xFF1558B5)) // QuizGameStyle.cardBackgroundColor
+                .padding(24.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Image(
+                bitmap = foxBitmap,
+                contentDescription = "Baron of Brainstorm",
+                modifier = Modifier.size(120.dp)
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            Text(
+                text = "It's Quiz Time!",
+                color = Color.White,
+                fontSize = 28.sp,
+                fontWeight = FontWeight.Bold,
+                textAlign = TextAlign.Center
+            )
+
+            Spacer(modifier = Modifier.height(12.dp))
+
+            Text(
+                text = "You think you're smart? The Baron of Brainstorm is ready to crush your ego. Are you?",
+                color = Color.White,
+                fontSize = 18.sp,
+                textAlign = TextAlign.Center
+            )
+
+            Spacer(modifier = Modifier.height(32.dp))
+
+            Box(
+                modifier = Modifier
+                    .fillMaxWidth(0.6f)
+                    .height(60.dp)
+                    .clip(RoundedCornerShape(12.dp))
+                    .background(Color(0xFF65B5FF)) // QuizGameStyle.defaultButtonColor
+                    .clickable { onReadyClick() },
+                contentAlignment = Alignment.Center
+            ) {
+                Text(
+                    text = "Ready",
+                    color = Color.White,
+                    fontSize = 18.sp,
+                    fontWeight = FontWeight.Bold
+                )
+            }
+        }
+    }
+}
+
+@Preview(showBackground = true, backgroundColor = 0xFF0B213E, device = "spec:parent=pixel_7, orientation=landscape")
+@Composable
+fun QuizReadyScreenPreview() {
+    QuizReadyScreen(
+        onReadyClick = {}
+    )
+}
